add auto publish to github nuget store
This includes:
- rename the package generated to not conflict with the one published to nuget.org
- Restructure the existing folder structure to ease deployment
- Add action to automatically publish to the nuget store on github
- Automatically generate version number
