+++ src/RenewedVision.Wpf.Interop.DirectX.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Wpf.Interop.DirectX_dxsdk", "Microsoft.Wpf.Interop.DirectX\Microsoft.Wpf.Interop.DirectX.csproj", "{51880B2F-A2DD-4669-B7F2-D5315B03C06D}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RenewedVision.Wpf.Interop.DirectX", "RenewedVision.Wpf.Interop.DirectX\RenewedVision.Wpf.Interop.DirectX.vcxproj", "{157A478D-FE02-4EB2-BD7C-8CF3BF1CB9A2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,14 +21,6 @@ Global
 		{157A478D-FE02-4EB2-BD7C-8CF3BF1CB9A2}.Release|x64.Build.0 = Release|x64
 		{157A478D-FE02-4EB2-BD7C-8CF3BF1CB9A2}.Release|x86.ActiveCfg = Release|Win32
 		{157A478D-FE02-4EB2-BD7C-8CF3BF1CB9A2}.Release|x86.Build.0 = Release|Win32
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Debug|x64.ActiveCfg = Debug|x64
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Debug|x64.Build.0 = Debug|x64
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Debug|x86.ActiveCfg = Debug|Win32
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Debug|x86.Build.0 = Debug|Win32
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Release|x64.ActiveCfg = Release|x64
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Release|x64.Build.0 = Release|x64
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Release|x86.ActiveCfg = Release|Win32
-		{51880B2F-A2DD-4669-B7F2-D5315B03C06D}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
