+++ src/RenewedVision.Wpf.Interop.DirectX/AssemblyInfo.cpp
