+++ src/RenewedVision.Wpf.Interop.DirectX/Stdafx.cpp
