+++ src/RenewedVision.Wpf.Interop.DirectX/Stdafx.h
