+++ src/RenewedVision.Wpf.Interop.DirectX/SurfaceDevice10.cpp
