+++ src/RenewedVision.Wpf.Interop.DirectX/SurfaceDevice9.cpp
