+++ src/RenewedVision.Wpf.Interop.DirectX/SurfaceQueue.cpp
