+++ src/RenewedVision.Wpf.Interop.DirectX/SurfaceQueue.inl
