+++ src/RenewedVision.Wpf.Interop.DirectX/SurfaceQueueImpl.h
