+++ src/RenewedVision.Wpf.Interop.DirectX/version.rc
